Projeto utilizado no curso SOLID com PHP: princípios da programação orientada a objetos
