--- SOLID-aula-5-completa/.gitignore
@@ -1,2 +1,0 @@
-.idea/
-vendor/
