--- SOLID-aula-5-completa/composer.json
@@ -1,8 +1,0 @@
-{
-  "name": "cviniciussdias/alura-solid",
-  "autoload": {
-    "psr-4": {
-      "Alura\\Solid\\": "src/"
-    }
-  }
-}
