--- SOLID-aula-5-completa/src/Model/AluraMais.php
@@ -1,24 +1,0 @@
-<?php
-
-namespace Alura\Solid\Model;
-
-class AluraMais extends Video implements Pontuavel
-{
-    private $categoria;
-
-    public function __construct(string $nome, string $categoria)
-    {
-        parent::__construct($nome);
-        $this->categoria = $categoria;
-    }
-
-    public function recuperarUrl(): string
-    {
-        return 'http://videos.alura.com.br/' . str_replace(' ', '-', strtolower($this->categoria));
-    }
-
-    public function recuperarPontuacao(): int
-    {
-        return $this->minutosDeDuracao() * 2;
-    }
-}
