--- SOLID-aula-5-completa/src/Model/Assistivel.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Alura\Solid\Model;
-
-interface Assistivel
-{
-    public function assistir(): void;
-}
