--- SOLID-aula-5-completa/src/Model/Feedback.php
@@ -1,29 +1,0 @@
-<?php
-
-namespace Alura\Solid\Model;
-
-class Feedback
-{
-    private $nota;
-    private $depoimento;
-
-    public function __construct(int $nota, ?string $depoimento)
-    {
-        if ($nota < 9 && empty($depoimento)) {
-            throw new \DomainException('Depoimento obrigatório');
-        }
-
-        $this->nota = $nota;
-        $this->depoimento = $depoimento;
-    }
-
-    public function recuperarNota(): int
-    {
-        return $this->nota;
-    }
-
-    public function recuperarDepoimento(): ?string
-    {
-        return $this->depoimento;
-    }
-}
