--- SOLID-aula-5-completa/src/Model/Pontuavel.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Alura\Solid\Model;
-
-interface Pontuavel
-{
-    public function recuperarPontuacao(): int;
-}
