--- SOLID-aula-5-completa/src/Service/Assistidor.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Alura\Solid\Service;
-
-use Alura\Solid\Model\Assistivel;
-
-class Assistidor
-{
-    public function assistir(Assistivel $conteudo)
-    {
-        $conteudo->assistir();
-    }
-}
