--- SOLID-aula-5-completa/src/Service/CalculadorPontuacao.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Alura\Solid\Service;
-
-use Alura\Solid\Model\Pontuavel;
-
-class CalculadorPontuacao
-{
-    public function recuperarPontuacao(Pontuavel $conteudo)
-    {
-        return $conteudo->recuperarPontuacao();
-    }
-}
