+++ php/solid-com-php/README.md
@@ -1,0 +1,13 @@
+# SOLID com PHP: princípios da programação orientada a objetos
+
+Categoria | Carga horária | Atividades | Finalizado | Certificado |
+:-:|:-:|:-:|:-:|:-:|
+Programação | 6 horas | 43 | 26/02/2022 | [Certificado Normal](https://cursos.alura.com.br/certificate/81bf4e64-0b53-48e9-a227-d17dde4839bd) / [Certificado Formal](https://cursos.alura.com.br/user/rodineicosta/course/solid-php-principios-orientacao-a-objetos/formalCertificate)
+
+## Aulas
+
+- Preparando o terreno
+- Melhorando a coesão
+- Trabalhando no acoplamento
+- Quebra de confiança
+- Encapsulando melhor
