+++ php/solid-com-php/composer.json
@@ -1,0 +1,8 @@
+{
+  "name": "cviniciussdias/alura-solid",
+  "autoload": {
+    "psr-4": {
+      "Alura\\Solid\\": "src/"
+    }
+  }
+}
