+++ php/solid-com-php/src/Model/Assistivel.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Alura\Solid\Model;
+
+interface Assistivel
+{
+    public function assistir(): void;
+}
