+++ php/solid-com-php/src/Model/Pontuavel.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Alura\Solid\Model;
+
+interface Pontuavel
+{
+    public function recuperarPontuacao(): int;
+}
