+++ php/solid-com-php/src/Service/Assistidor.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Alura\Solid\Service;
+
+use Alura\Solid\Model\Assistivel;
+
+class Assistidor
+{
+    public function assistir(Assistivel $conteudo)
+    {
+        $conteudo->assistir();
+    }
+}
