+++ php/solid-com-php/src/Service/CalculadorPontuacao.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Alura\Solid\Service;
+
+use Alura\Solid\Model\Pontuavel;
+
+class CalculadorPontuacao
+{
+    public function recuperarPontuacao(Pontuavel $conteudo)
+    {
+        return $conteudo->recuperarPontuacao();
+    }
+}
